requirements.txt
# --- Основные библиотеки ---
pydantic>=2.0          # Модели данных и валидация JSON
python-docx>=1.0.0     # Генерация и экспорт DOCX файлов
python-dotenv>=1.0.0   # Загрузка конфигов из .env

# --- Для асинхронной работы с API (если планируешь подключать AI) ---
httpx>=0.27.0          # Асинхронные HTTP-запросы к API моделей

# --- Удобное логирование (по желанию) ---
loguru>=0.7.2          # Красивые логи, удобно для отладки

# --- Дополнительно ---
# aiofiles>=23.2.1     # Если захочешь асинхронно работать с файлами
g4f
tqdm